fixed two breaks caused by changing SR.experimentalWellCount from int to Integer
git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1966 80c535d0-b110-0410-b4de-9279f94142c3

+++ test/edu/harvard/med/screensaver/db/ComplexDAOTest.java
@@ -847,7 +847,7 @@ public class ComplexDAOTest extends AbstractSpringTest
         Screen screen = genericEntityDao.findEntityByProperty(Screen.class, "screenNumber", 1);
         assertEquals("plate numbers", expectedPlateNumbers, screen.getScreenResult().getPlateNumbers());
         assertEquals("wells", expectedWells, screen.getScreenResult().getWells());
-        assertEquals("experimental well count", expectedExperimentalWellCount[0], screen.getScreenResult().getExperimentalWellCount());
+        assertEquals("experimental well count", expectedExperimentalWellCount[0], (int) screen.getScreenResult().getExperimentalWellCount());
       }
     });
   }
